requirements.txt
fastapi
uvicorn
kafka-python
pydantic
mlflow
pyspark
boto3
